Use floats instead of scaled ints for PR score (#872)
* Update PR Benchmark

* Use float instead of scaled int to keep track of page ranks

* Combine PR scores in a single pass, same as the HugePR does

* Fix LongIterators benchmark

* Ignore daming factor in HugeEigenvector PR

+++ algo/src/main/java/org/neo4j/graphalgo/impl/pagerank/HugeComputeStep.java
@@ -23,9 +23,9 @@ public interface HugeComputeStep extends Runnable {
 
     long[] starts();
 
-    void prepareNextIteration(int[][] score);
+    void prepareNextIteration(float[][] score);
 
-    int[][] nextScores();
+    float[][] nextScores();
 
     void setStarts(long[] startArray, int[] lengthArray);
